SS4 compatibility: Implement core namespaces, update requirements
