+++ change-log.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 2.0.0 (unreleased)
+
+* SilverStripe 4 minimum required
+* PHP 5.5 minimum required
+
 ## 1.0.0
 
 * Initial release
